requirements.txt
flask==2.0.1
pymongo==4.3.3
bcrypt==3.2.0
python-dotenv==0.19.0
google-generativeai==0.1.0
opencv-python==4.5.5.64
fer==22.3.0
pandas==1.3.5
requests==2.26.0
gunicorn==20.1.0